add a cast from size_t to int64_t

+++ sdk/core/azure-core/src/io/random_access_file_body_stream.cpp
@@ -37,7 +37,7 @@ int64_t RandomAccessFileBodyStream::OnRead(
   auto numberOfBytesRead = pread(
       this->m_fileDescriptor,
       buffer,
-      std::min(count, this->m_length - this->m_offset),
+      std::min(static_cast<int64_t>(count), this->m_length - this->m_offset),
       this->m_baseOffset + this->m_offset);
 
   if (numberOfBytesRead < 0)
